Fix broken S3 integration (#7368)
## Summary of changes

The latest version of S3 is broken

## Reason for change

This is the same issue we've seen several times, the s3 integration is
broken, because we're ducktyping the field (which changes) instead of
the public property

## Implementation details

Update _all_ the S3 integrations to switch to using properties instead
of the backing field, because it's just a matter of time before they all
break (I manually checked they're all valid).

Also fixed the nullability, which was incorrect (they can all
technically be null).

And fixed potential null ref issue because of that related to span
pointers.

## Test coverage

This is the test (I bumped the package versions) but will also did an
all-versions test that passed for all the versions and all TFMs

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->