Include `telemetry_forwarder.exe` in the Windows SSI OCI images (#7342)
## Summary of changes

Downloads and verifies the telemetry_forwarder.exe binary and includes
it in the OCI image

## Reason for change

We want to ship the binary side-by-side with the fleet installer. At
some point we'll configure a proper release process for the
`telemetry_forwarder.exe` dll, but in the mean time we're simply pulling
from a public location for simplicity.

## Implementation details

- Download the telemetry forwarder from our Azure blob storage
- Sign the binary as part of our normal GitLab build
- Include the forwarder in our GitLab build output
- Set the `DD_TELEMETRY_FORWARDER_PATH` when installing the tracer

## Test coverage

Less than I'd like, but ultimately this is a non-critical component,
that is tested independently. It would be nice to have some testing that
the telemetry is forwarded correctly, but ultimately I think this should
be added in the system-tests/onboarding tests repo. I included the
telemetry forwarder in our smoke tests so we should at least get
warnings/errors there if there's anything egregious.

## Other details

Stacked on 
- https://github.com/DataDog/dd-trace-dotnet/pull/7341