Fix more flake in the IIS fleet smoke tests (#7372)
## Summary of changes

Fixes more flake in the IIS smoke tests

## Reason for change

A recent WAF rules update #7331 started adding more tags to spans, which
ultimately can cause flake. Additionally, there are other tags which are
only added to the very first span generated, but in a race-y way, so
that can cause flake

## Implementation details

Add more exclusions to the smoke tests in general. I signed off on these
exclusions with ASM (they suggested it!)

## Test coverage

If the tests pass, we're good

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->