Ignore rejit timeout errors in smoke tests and report a metric (#7370)
## Summary of changes

Ignore the "rejit timed out" errors in smoke tests, but report a metric

## Reason for change

We're still seeing flake from this. We don't want to increase the
timeout further at this stage, so instead we want to allow it, and
report a metric so that we can track the issue

## Implementation details

Added a helper for sending a metric, and added a "reportablePatterns"
list for the smoke tests.

## Test coverage

This is the test, will see if any of the jobs hit it and 🤞 

## Other details

We will set up a dashboard to monitor these (and other skip metrics) as
part of the pre-release process