[CI] Creates PR Label Based Docker Image (#7337)
## Summary of changes
Adding new stage to check if a label has been added to the PR for Azure
to decide wether a Docker image should be created based on the PR or
not(currently we only do it for master after each merge).

- Related system test PR where I tested the image being used correctly(a
4 lines change):
https://github.com/DataDog/system-tests/pull/5024
- Location where the image gets used and all tests pass: 

https://github.com/DataDog/system-tests/actions/runs/16763951714/job/47465265591?pr=5024#step:7:72

## Reason for change
For us to be able to review/merge system-test PRs with higher confidence
and remove the need to run locally for cases where the user may not
want/need to.

## Implementation details
Updated the action, workflow and pipeline files needed to take into
account the label and use the branch name instead.

## Test coverage
This is the test, the first commits add the docker image and tag it as
expected:
[Location](https://github.com/DataDog/dd-trace-dotnet/pkgs/container/dd-trace-dotnet%2Fdd-trace-dotnet)
The last commit makes sure the image gets deleted when this PR(with its
label) closes, so the above should not have it anymore:
[Location](https://github.com/DataDog/dd-trace-dotnet/pkgs/container/dd-trace-dotnet%2Fdd-trace-dotnet)

There could still be a scenario where the label gets removed and then
the PR closed where the image could still remain but I doubt any of us
would that...