Maximo/label based artifacts (#7385)
## Summary of changes
Left the wrong name in the PR for the master based snapshots, so fixing
it to be `latest_snapshot` instead of `master`.
Updating the delete method since the current one didn't delete the image
as expected.

## Reason for change
Created the wrong package under: 
-
https://github.com/DataDog/dd-trace-dotnet/pkgs/container/dd-trace-dotnet%2Fdd-trace-dotnet
<img width="598" height="476" alt="image"
src="https://github.com/user-attachments/assets/46e7e114-852b-4409-b888-c84d9f89a970"
/>

## Implementation details

## Test coverage
This PR is part of the change.