Fix Windows SSI release (#7379)
## Summary of changes

Fixes the Windows SSI release

## Reason for change

In the last release, 3.24.0, the SSI images failed because the Windows
OCI image could not be built. This is due to inherent differences
between how the artifacts are handled for the release versus normal
builds (and means we can't actually _test_ the "release" path until we,
you know, release).

## Implementation details

Make sure we publish the `telemetry_forwarder.exe` as part of the
publish stage, and then pull this when creating the release.

## Test coverage

It worked in the 3.24.1 release, so we're good now

## Other details

It really feels like this whole process is kind of broken. The OCI
images should be created _once_ and then just promoted as part of the
release, not completely re-built, but that's on the libdatadog-build
team...