Disable debugger initialization if it's not enabled initially (#7398)
## Summary of changes

Disables the debugger initialization if it's not explicitly enabled

## Reason for change

We have been seeing some hangs in CI which seem to be caused by the
debugger (based on log messages) and I suspect are due to [this
refactoring](https://github.com/DataDog/dd-trace-dotnet/commit/015d21801674f10d8d72c17276723ec9c322df69).
Rather than revert that large piece of refactoring (and causing massive
conflicts for the [remote-enablement
PR](https://github.com/DataDog/dd-trace-dotnet/pull/7366)), for now we
primarily want to ensure customers _not_ using the debugger are not
impacted.

## Implementation details

Check the debugger and exception replay settings to see if the debugger
is required. If not, bail out of initialization immediately.

## Test coverage

This is the test