Add final job for blocking merges of PRs (#7376)
## Summary of changes

Adds a "final" job at the end of ultimate-pipeline that can be used to
block merges

## Reason for change

We have had various cases where the PR had failed, but we merged anyway,
and that broke master. A big reason for that was flakiness, but as we
continue to work on that, we want to make it easy to require a full
green pipeline before merges. Unfortunately, that's a PITA with Azure
DevOps 😅

## Implementation details

- Add two new stages which depend on every other stage, but have
conditions such that they run even if previous stages didn't run
- When those stages run, post the results to GitHub
- Create a job to auto-generate those files (because they're
super-verbose)

## Test coverage

Ran this a few times. Confirmed that if anything fails, this job fails.
If everything passes (or didn't run), this job passes.

## Other details

We can later decide when we want to require this job. Note that we
should _also_ still require _some_ jobs that we expect to always run, to
make sure that things have run that we expect to


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->