Add more alpine checks for native library (#7395)
## Summary of changes

Refactors the native snapshot checks and adds explicit validation of the
symbols used by the native libraries against an allowlist

## Reason for change

We are shortly going to start building with `alpine:3.17` instead of
`alpine:3.14`, but we don't want to break customers, so we need to make
sure we only use symbols that are available in the earlier version of
musl.

We previously added snapshots for which symbols we depend on here:
- https://github.com/DataDog/dd-trace-dotnet/pull/7224

That means we can see exactly when we depend on new symbols.

However, we realised that libdatadog is already building with a much
later version of alpine (3.22?) and so wanted to ensure that we're not
_Already_ depending on symbols that are not available in some code
paths.

## Implementation details

There are two parts to this PR

- Extract the existing symbol snapshotting out of `Build.cs` and into a
helper class. There's some minor renaming etc, but otherwise this is
just copy-paste, so you can skip the first commit.
- Extract the symbols defined by musl 1.2.2 and libgcc (which the
profiler/libdatadog use for libunwind stuff). These form the "allow
list" of symbols (libgcc is allowed [because .NET depends on it
too](https://learn.microsoft.com/en-us/dotnet/core/install/linux-alpine?tabs=dotnet9#dependencies)).

Now when we extract the symbols for alpine/universal binaries, we check
that we are only depending on symbols defined in these libraries.

> Note that we ignore `ddog_` and `blaze_` symbols which are defined in
libdatadog and required by the profiler, on the assumption that there
will never be a mismatch here.

## Test coverage

Tested locally with all the native libraries we produce, and all looks
good, so as long as CI passes we're good.

## Other details

We should probably add snapshotting of the WAF too to be safe 🤔