Scrape native tracer stats from logs in smoke tests and track in Datadog (#7410)
## Summary of changes

Scrapes the stats that we write from the native tracer in to the logs
and sends them to Datadog

## Reason for change

We aren't currently collecting these stats, and it would be good to
understand the distribution of values we get, how they change over time,
and how they vary per runtime.

## Implementation details

- Parse all the native logs looking for the stats line
- Parse the stats out
- Send metrics:
- `dd_trace_dotnet.ci.smoke_tests.init_duration_ms`
- `dd_trace_dotnet.ci.smoke_tests.callback_duration_ms`
- `dd_trace_dotnet.ci.smoke_tests.callback_count`
- `dd_trace_dotnet.ci.smoke_tests.callback_mean_duration_ms`

Where the latter three are tagged with
- `native:initialize`
- `native:module_load_finished`
- `native:calltarget_request_rejit`
- `native:calltarget_request_rewriter`
- `native:assembly_load_finished`
- `native:module_unload_started`
- `native:jit_compilation_started`
- `native:jit_inlining`
- `native:jit_cache_function_search_started`
- `native:initialize_profiler`
- `native:enquue_rejit_for_loaded_modules`

## Test coverage

Tested locally that the scraping works etc

## Other details

Requires https://github.com/DataDog/dd-trace-dotnet/pull/7409 for
correct tagging:
- https://github.com/DataDog/dd-trace-dotnet/pull/7409


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->