Add missing `GetDictionary()` method to `ConfigurationBuilder` (#7421)
## Summary of changes

Add missing `GetDictionary()` method to `ConfigurationBuilder`

## Reason for change

We were missing a method that allows providing a default instance for a
dictionary up-front. This was forcing using other overloads that are
suboptimal for telemetry.

## Implementation details

Add the missing helper that allows providing a default value. I allowed
passing a null value in here, as it avoids unnecessary allocations if
you don't actually need the dictionary. A good example was the fixed
usage in `InitializeHeaderTags` which was creating an empty dictionary
instead of taking the fast-path `null` route (and potentially actually
giving _incorrect_ settings in the `DynamicSettings` I think

## Test coverage

Meh

## Other details

Spotted when reviewing
- https://github.com/DataDog/dd-trace-dotnet/pull/7420


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->