Add DD_LOGS_DIRECT_SUBMISSION_AZURE_FUNCTIONS_HOST_ENABLED
Allows for the ability to disable direct log
submission when we detect that we are running on
the function host.

When using direct log submission in a dotnet-isolated
Azure Functions application we end up getting the
worker logs re-logged by the host process which
makes it appear that there are duplicated logs.

This allows for the ability to entirely disable
direct log submission on that function host.