Revert "[Profiler] Make `timer_create`-based CPU profiler default (#7322)" (#7427)
## Summary of changes

This reverts commit 5676931004042cfb4af0468567f00c056d2d467f.

## Reason for change

We noticed a lot of (flaky) crashing on the _trimmed_ smoke tests,
shortly after #7322 was merged. Investigating this by testing a revert
of the PR and re-running in CI resulted in no crashes. Out of an
abundance of caution, we are choosing to revert the change to
investigate before the next release.

## Implementation details

This reverts commit 5676931004042cfb4af0468567f00c056d2d467f 
- #7322

## Test coverage

This is the test, but I also re-ran all the trimmed tests 6 times and
saw no failures

## Other details

We'll aim to revert this revert and investigate next week


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->