[Tracer] fix: Re-use runtime metrics writer resources to limit memory growth (#7434)
## Summary of changes
Updates `TracerManagerFactory.CreateTracerManager` to pass and re-use
the previous `RuntimeMetricsWriter`, if runtime metrics is enabled for
the new TracerManager. We still create a new `IDogStatsd` client every
time, so any updated Agent settings are adopted by the DogStatsd client,
but this could be further optimized at a later time to re-use the client
if none of its configuration (e.g. host/port/tags) have changed.

## Reason for change
We've observed a scenario where the number of `RuntimeEventListener`
instances continues to grow, consuming more and more memory. This
happens whenever new Dynamic Configuration settings are received by the
tracer and runtime metrics are enabled. This PR resolves the issue.

## Implementation details
When creating the new `TracerManager`, pass in the previous
`RuntimeMetricsWriter` instance and only update the `IDogStatsd` object
with new settings. This makes sure that we maintain only one
`RuntimeMetricsWriter` instance while getting up-to-date DogStatsD
settings throughout the application lifetime.

## Test coverage
Adds a small unit test to confirm that the previous
`RuntimeMetricsWriter` is re-used. Additionally, local testing was done
to confirm that the number of
`Datadog.Trace.RuntimeMetrics.RuntimeEventListener` objects does not
grow when Dynamic Configuration is updated.

### Without the fix
After some number of Dynamic Configuration settings were made in the
Datadog UI, a dump was taken with the following analysis:
```
> dumpheap -type Datadog.Trace.RuntimeMetrics.RuntimeEventListener
         Address               MT           Size
    0159c3c7fe88     7ffe4e801810             88
    0159c442ed28     7ffe4e801810             88
    0159c44eac60     7ffe4e801810             88
    0159c4500f00     7ffe4e801810             88
    0159c454ea68     7ffe4e801810             88

Statistics:
          MT Count TotalSize Class Name
7ffe4e801810     5       440 Datadog.Trace.RuntimeMetrics.RuntimeEventListener
Total 5 objects, 440 bytes
```

Then after one additional Dynamic Configuration update was made in the
Datadog UI, a dump was taken with the following analysis:
```
> dumpheap -type Datadog.Trace.RuntimeMetrics.RuntimeEventListener
         Address               MT           Size
    0159c3c7fe88     7ffe4e801810             88
    0159c442ed28     7ffe4e801810             88
    0159c44eac60     7ffe4e801810             88
    0159c4500f00     7ffe4e801810             88
    0159c454ea68     7ffe4e801810             88
    0159c61fef48     7ffe4e801810             88

Statistics:
          MT Count TotalSize Class Name
7ffe4e801810     6       528 Datadog.Trace.RuntimeMetrics.RuntimeEventListener
Total 6 objects, 528 bytes
```

### With the fix
After some number of Dynamic Configuration settings were made in the
Datadog UI, a dump was taken with the following analysis:
```
> dumpheap -type Datadog.Trace.RuntimeMetrics.RuntimeEventListener
         Address               MT           Size
    01527107faf0     7ffe4e7acc00             88

Statistics:
          MT Count TotalSize Class Name
7ffe4e7acc00     1        88 Datadog.Trace.RuntimeMetrics.RuntimeEventListener
Total 1 objects, 88 bytes
```

Then after one additional Dynamic Configuration update was made in the
Datadog UI, a dump was taken with the following analysis:
```
> dumpheap -type Datadog.Trace.RuntimeMetrics.RuntimeEventListener
         Address               MT           Size
    01527107faf0     7ffe4e7acc00             88

Statistics:
          MT Count TotalSize Class Name
7ffe4e7acc00     1        88 Datadog.Trace.RuntimeMetrics.RuntimeEventListener
Total 1 objects, 88 bytes
```

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>