Use the global.json file to specify the .NET SDK version in GitHub Actions (#7442)
## Summary of changes

Point the `setup-dotnet` action at the global.json

## Reason for change

Means we don't need to update these when we change the version of the
SDK we use

## Implementation details

Find and replace

## Test coverage

If this works, I think we're ok

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->