Update Managed DevOps pools images (#7446)
## Summary of changes

Updates the VM images to be based on Ubuntu 24.04 instead of 20.04

## Reason for change

We've been plagued by `##[error]The remote provider was unable to
process the request.` errors which cause jobs to just never start. After
a call with Microsoft, they suggested the issue was related to using
Ubuntu 20.04 as our base image for the VMs. Updating the images to test
it out.

## Implementation details

Rebuild the Linux VMs to use the new base images. The bulk of the
implementation happens in dockerfiles so hopefully there won't be any
issues

## Test coverage

This is the test

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->