[Native] Convert Rejitters vector to a fixed array (#7460)
## Summary of changes
After converting the creation of Dataflow to dynamic, when adding it to
the rejiters vector, this could cause a race condition when it grows.
So, in order to avoid this, the rejitters vector has been converted to a
fixed array.

As there are 3 known rejitters, ATM, the size of the array has been set
to 4 to have a minimal buffer.

## Reason for change
Previous vector reserve implementation could be unsafe if a new rejitter
was added without modifying the reserved size. This way if a new one is
added without modifying the max count, an exception will be thrown

## Implementation details
Replace the previous rejitters vector with a fixed array.

## Test coverage
Unit tests have been added for the array insert function

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->