Consolidate bailout checks into native loader (#7462)
## Summary of changes

Move some of the bail-out checks from the native loader into the native
tracer

## Reason for change

As far as I can tell, these checks really _should_ be in the native
loader, and they're only in the native tracer currently for historical
reasons. This results in both duplication of work _and_ mismatch between
traced processes for the profiler and the tracer. This was discovered as
part of the work in #7287.

## Implementation details

- Removed duplicate checks (process name variables) from the native
tracer
- Moved the additional CI Visibility checks to the native loader. I'll
defer to @tonyredondo if we want to tidy these up at all
- Moved the AAS checks to the native loader. I'll defer to
@lucaspimentel if we want to change any of these.

## Test coverage

This should be covered by the existing tests in general, but I may add
an additional instrumentation test to make sure we get no tracer logs
for previous cases where we would have done.

## Other details

Currently the native tracer prints a bunch of env vars for debugging
purposes. We could move this to the native loader too if we prefer.


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>