Update language used in native log messages (#7467)
## Summary of changes

Updates some of the language used in native logs to avoid ambiguous use
of `Profiler`

## Reason for change

`Profiler` is tricky. We use the .NET profiling APIs for all our native
functionality, so our native components are technically all "profilers".
But this is ambiguous with the Continuous Profiler _product_, and can
often confuse both customers and support.

To reduce confusion, we should aim to _avoid_ using "Profiler" to refer
to the native components where possible. In particular

- Inside the native loader, when we disable the whole product, use
`Datadog SDK`
  - e.g. `The Datadog SDK has been disabled`
- Always refer to the Continuous Profiler as `Continuous Profiler`
  - e.g. `The Continuous Profiler has been disabled`
- Refer to the native tracer as either `Instrumentation` or
`Instrumentation component` as appropriate
- e.g. `Instrumentation has been disabled` or `The Instrumentation
component failed to initialize`
- Don't refer to the "managed profiler", instead use `Datadog.Trace.dll`
(I hope it's not necessary to refer to the
Datadog.Trace.ClrProfiler.Managed.Loader.dll 😅)
- e.g. `Unable to initialize: Datadog.Trace.dll was not yet loaded into
the App Domain`


Note that we're _primarily_ concerned with the "high level" log messages
which are "customer facing". See other details for exceptions

## Implementation details

Updated the logs in the native loader and native tracer in line with
this guidance. I'm assuming the CP already follows this.

## Test coverage

N/A

## Other details

Note that internal/technical naming is still valid, so we still have the
same distinct components, e.g.

- Native loader
- Native tracer
- Managed tracer loader
- Managed tracer
- Libdatadog
- Continuous Profiler

Additionally, we can still use `CorProfiler` / `ICorProfiler` / `COR
Profiler` to refer to the specific runtime components if necessary.

---------

Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>
Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>