Improve logging for `InstrumentationTests` and update dynamic code scenario (#7472)
## Summary of changes

- Add some missing `SetLogDirectory()` calls for `InstrumentationTests`
- Update the "dynamic code" instrumentation tests to throw away the logs

## Reason for change

`InstrumentationTests` tests a bunch of very different scenarios. To
make it easier to understand the failures, we generally put the logs for
each test into its own subfolder by calling `SetLogDirectory()`. We were
missing this call for some tests, so added them.

Additionally, while working on #7287 we found a scenario which is
currently "broken"
- We instrument an app that has "dynamic code disabled"
- We don't support this, so we bail out in the _managed loader_
- The `Tracer` is never initialized
- The Continuous Profiler writes an error on shutdown because the Tracer
never sends the stable config.
- In `CheckBuildLogsForErrors` we fail the CI because of the error log.

_Ideally_ we would bail out in the _native_ loader so that nothing's
loaded, but we don't know how to read the context switch on the native
side ATM, so we can't. However, as this isn't a supported scenario,
we're ok with the error log being there, and don't think it's worth
trying to avoid it at this stage.

## Implementation details

- Add missing `SetLogDirectory()`
- Write the "dynamic code" logs to a "temp" directory so that they're
not checked by the `CheckBuildLogsForErrors` stage

## Test coverage

Coverage is the same

## Other details

Blocker for 
- https://github.com/DataDog/dd-trace-dotnet/pull/7287